+++ src/zrb/session_state_logger/any_session_state_logger.py
@@ -1,7 +1,10 @@
 import datetime
 from abc import ABC, abstractmethod
 
-from zrb.session_state_log.session_state_log import SessionStateLog, SessionStateLogList
+from zrb.session_state_log.session_state_log_model import (
+    SessionStateLog,
+    SessionStateLogList,
+)
 
 
 class AnySessionStateLogger(ABC):
