+++ src/zrb/util/file.py
@@ -32,8 +32,10 @@ def _read_text_file_content(file_path: str) -> str:
 
 def _read_pdf_file_content(file_path: str) -> str:
     import pdfplumber
+    from pdfplumber.pdf import PDF
 
     with pdfplumber.open(file_path) as pdf:
+        pdf: PDF
         return "\n".join(
             page.extract_text() for page in pdf.pages if page.extract_text()
         )
